requirements.txt
python-pptx
langgraph
langchain-core
fastapi
uvicorn
python-multipart
faiss-cpu
aiohttp
tiktoken
openai
python-dotenv
openpyxl
matplotlib
numpy
beautifulsoup4
requests
playwright>=1.41.0
pydantic>=2.5.0
google-genai>=0.4.0
python-docx>=1.1.0
